+++ TransitionTreasuryTabbarDemo/AppDelegate.swift
@@ -2,7 +2,7 @@
 //  AppDelegate.swift
 //  TransitionTreasuryTabbarDemo
 //
-//  Created by 宋宋 on 16/1/16.
+//  Created by DianQK on 16/1/16.
 //  Copyright © 2016年 TransitionTreasury. All rights reserved.
 //